Refactor History Module: split files; use confirmation dialogs * Decompose HistoryRootView (919 LOC) into per-mode extension files   (Treatments, Meals, Glucose, Adjustments) plus Filters, AddGlucose,   and Confirmations. * Replace three duplicated .alert(...) modifiers (all sharing one   buggy isRemoveHistoryItemAlertPresented boolean) with a single   .confirmationDialog driven by a unified History.DeletionTarget   enum. * Add deleteInsulinFromTidepool and deleteCarbsFromTidepool to the   HistoryProvider protocol (previously called only via the concrete   Provider type). * Remove confirmed dead code: logGlucoseButton, filterEntriesButton,   deleteGlucose(at:), filterPopoverAnchor.
